destroy the memory leak
