+++ src/pages/Staking/components/InfoBlock/InfoBlock.js
@@ -1,11 +1,14 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import * as PropTypes from 'prop-types';
 import { ReactSVG } from 'react-svg';
 import ReactTooltip from 'react-tooltip';
+import { BigNumber } from 'ethers';
 
 import Paragraph from '../../../../components/Paragraph';
-import { Loader } from '../../../../components/Loader';
-import { formatRounded } from '../../../../services/staking.wrapper';
+import {
+  FIXED_POINT,
+  formatRounded,
+} from '../../../../services/staking.wrapper';
 import earningsIcon from '../../../../assets/svg/last24h.svg';
 import pieChartOutlineIcon from '../../../../assets/svg/pie_chart_outline.svg';
 import errorOutlineIcon from '../../../../assets/svg/error_outline.svg';
@@ -13,171 +16,212 @@ import copyIcon from '../../../../assets/svg/copy.svg';
 import useCopyToClipboard from '../../../../hooks/useCopyToClipboard';
 import DisplayValue from '../../../../components/DisplayValue';
 import { tooltips } from '../../../../config';
+import appStore from '../../../../store/app.store';
 
-const InfoBlock = ({
-  account,
-  totalReward,
-  totalRewardInUsd,
-  totalStakedInUsd,
-  totalStaked,
-}) => {
+const InfoBlock = ({ poolsArr, account }) => {
   const { isCopied, onCopy } = useCopyToClipboard({ text: account });
+  const [totalReward, setTotalReward] = useState(null);
+  const [totalStaked, setTotalStaked] = useState(null);
+  const [totalRewardInUsd, setTotalRewardInUsd] = useState(null);
+  const [totalStakedInUsd, setTotalStakedInUsd] = useState(null);
 
+  useEffect(() => {
+    (async () => {
+      if (appStore.stakingWrapper !== undefined && poolsArr.length > 0) {
+        const poolsRewards = [];
+        const myTotalStake = [];
+        poolsArr.map(async (pool) => {
+          const { estAR, myStakeInAMB } =
+            await appStore.stakingWrapper.getPoolData(pool.index);
+          if (estAR) {
+            poolsRewards.push(estAR);
+            const rewardInAmb =
+              poolsRewards.length > 0 &&
+              poolsRewards.reduceRight((acc, curr) => acc + +curr, 0);
+            setTotalReward(rewardInAmb > 0 && rewardInAmb);
+            const esdSum =
+              appStore.tokenPrice &&
+              poolsRewards.length > 0 &&
+              poolsRewards.reduceRight((acc, curr) => acc + +curr, 0);
+            setTotalRewardInUsd(
+              esdSum &&
+                appStore.tokenPrice &&
+                esdSum > 0 &&
+                esdSum * appStore.tokenPrice,
+            );
+          }
+          if (myStakeInAMB) {
+            myTotalStake.push(myStakeInAMB);
+            if (myTotalStake.length > 0) {
+              const totalStakeSum = myTotalStake.reduceRight(
+                (acc, curr) => acc.add(curr),
+                BigNumber.from('0'),
+              );
+              setTotalStaked(
+                totalStakeSum &&
+                  totalStakeSum.gte(FIXED_POINT) &&
+                  totalStakeSum,
+              );
+              setTotalStakedInUsd(
+                totalStakeSum &&
+                  appStore.tokenPrice &&
+                  totalStakeSum.gte(FIXED_POINT) &&
+                  formatRounded(totalStakeSum) * appStore.tokenPrice,
+              );
+            }
+          }
+        });
+      }
+    })();
+  }, [appStore.refresh, account]);
   return (
     <div className="info-block ">
       <div className="wrapper">
-        {account ? (
-          <>
-            <div className="info-block__address">
-              <Paragraph size="m-400" style={{ paddingBottom: 5 }}>
-                My Address
-              </Paragraph>
-              <Paragraph
-                size="l-500"
-                style={{ color: '#333333', fontSize: 16, fontWeight: 400 }}
-              >
-                {account
-                  ? ` ${account.substr(0, 9)}...${account.slice(32)}`
-                  : '...'}{' '}
-                <ReactSVG
-                  data-tip
-                  data-for="copy-state"
-                  onClick={onCopy}
-                  src={copyIcon}
-                  wrapper="span"
-                  style={{ marginLeft: 20, cursor: 'pointer' }}
-                />
-              </Paragraph>
-              {!isCopied ? (
-                <ReactTooltip id="copy-state" place="top" effect="solid">
-                  {tooltips.copyState.notCopied}
+        <>
+          <div className="info-block__address">
+            <Paragraph size="m-400" style={{ paddingBottom: 5 }}>
+              My Address
+            </Paragraph>
+            <Paragraph
+              size="l-500"
+              style={{ color: '#333333', fontSize: 16, fontWeight: 400 }}
+            >
+              {account
+                ? ` ${account.substr(0, 9)}...${account.slice(32)}`
+                : '...'}{' '}
+              <ReactSVG
+                data-tip
+                data-for="copy-state"
+                onClick={onCopy}
+                src={copyIcon}
+                wrapper="span"
+                style={{ marginLeft: 20, cursor: 'pointer' }}
+              />
+            </Paragraph>
+            {!isCopied ? (
+              <ReactTooltip id="copy-state" place="top" effect="solid">
+                {tooltips.copyState.notCopied}
+              </ReactTooltip>
+            ) : (
+              <ReactTooltip id="copy-state" place="top" effect="solid">
+                {tooltips.copyState.isCopied}
+              </ReactTooltip>
+            )}
+          </div>
+          <div className="info-block__stacked">
+            <div className="info-block__stacked--total">
+              <div>
+                <ReactTooltip id="total-staked" place="top" effect="solid">
+                  {tooltips.totalStaked}
                 </ReactTooltip>
-              ) : (
-                <ReactTooltip id="copy-state" place="top" effect="solid">
-                  {tooltips.copyState.isCopied}
-                </ReactTooltip>
-              )}
-            </div>
-            <div className="info-block__stacked">
-              <div className="info-block__stacked--total">
-                <div>
-                  <ReactTooltip id="total-staked" place="top" effect="solid">
-                    {tooltips.totalStaked}
-                  </ReactTooltip>
+                <div style={{ display: 'flex', flexDirection: 'row' }}>
                   <div style={{ display: 'flex', flexDirection: 'row' }}>
-                    <div style={{ display: 'flex', flexDirection: 'row' }}>
-                      <ReactSVG
-                        style={{
-                          paddingTop: 1,
-                        }}
-                        src={pieChartOutlineIcon}
-                        wrapper="span"
-                      />
-                      <Paragraph
-                        size="m-400"
-                        style={{ paddingBottom: 5, whiteSpace: 'nowrap' }}
-                      >
-                        &nbsp;Holdings&nbsp;
-                      </Paragraph>
-                    </div>
                     <ReactSVG
-                      data-tip
-                      data-for="total-staked"
-                      src={errorOutlineIcon}
+                      style={{
+                        paddingTop: 1,
+                      }}
+                      src={pieChartOutlineIcon}
                       wrapper="span"
                     />
+                    <Paragraph
+                      size="m-400"
+                      style={{ paddingBottom: 5, whiteSpace: 'nowrap' }}
+                    >
+                      &nbsp;Holdings&nbsp;
+                    </Paragraph>
                   </div>
-                </div>
-                <div
-                  style={{
-                    minWidth: '100%',
-                    display: 'flex',
-                    flexDirection: 'row',
-                    justifyContent: 'space-around',
-                  }}
-                >
-                  <DisplayValue
-                    size="xl-400"
-                    color="#1ACD8C"
-                    value={totalStaked && formatRounded(totalStaked, 2)}
-                  />
-                  <Paragraph size="xl-400" style={{ color: '#4A38AE' }}>
-                    &nbsp; / &nbsp;
-                  </Paragraph>
-                  <DisplayValue
-                    size="xl-400"
-                    color="#4A38AE"
-                    value={totalStakedInUsd}
-                    symbol="$"
-                  />
-                </div>
-              </div>
-              <div className="info-block__stacked--course">
-                <div
-                  style={{
-                    display: 'inline-flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                  }}
-                >
-                  <ReactTooltip id="earnings" place="top" effect="solid">
-                    {tooltips.earnings}
-                  </ReactTooltip>
-                  <ReactSVG
-                    style={{
-                      paddingTop: 0,
-                    }}
-                    src={earningsIcon}
-                  />
-                  <Paragraph size="m-400" style={{ paddingBottom: 5 }}>
-                    &nbsp;Est. yearly yield&nbsp;
-                  </Paragraph>
                   <ReactSVG
                     data-tip
-                    data-for="earnings"
+                    data-for="total-staked"
                     src={errorOutlineIcon}
                     wrapper="span"
                   />
                 </div>
-
-                <div
-                  style={{
-                    minWidth: '100%',
-                    display: 'flex',
-                    flexDirection: 'row',
-                    justifyContent: 'space-around',
-                  }}
-                >
-                  <DisplayValue
-                    size="xl-400"
-                    color="#1ACD8C"
-                    value={totalReward}
-                  />
-                  <Paragraph size="xl-400" style={{ color: '#4A38AE' }}>
-                    &nbsp; / &nbsp;
-                  </Paragraph>
-                  <DisplayValue
-                    size="xl-400"
-                    color="#4A38AE"
-                    value={totalRewardInUsd}
-                    symbol="$"
-                  />
-                </div>
+              </div>
+              <div
+                style={{
+                  minWidth: '100%',
+                  display: 'flex',
+                  flexDirection: 'row',
+                  justifyContent: 'space-around',
+                }}
+              >
+                <DisplayValue
+                  size="xl-400"
+                  color="#1ACD8C"
+                  value={totalStaked && formatRounded(totalStaked, 2)}
+                />
+                <Paragraph size="xl-400" style={{ color: '#4A38AE' }}>
+                  &nbsp; / &nbsp;
+                </Paragraph>
+                <DisplayValue
+                  size="xl-400"
+                  color="#4A38AE"
+                  value={totalStakedInUsd}
+                  symbol="$"
+                />
               </div>
             </div>
-          </>
-        ) : (
-          <Loader types="spokes" />
-        )}
+            <div className="info-block__stacked--course">
+              <div
+                style={{
+                  display: 'inline-flex',
+                  alignItems: 'center',
+                  justifyContent: 'center',
+                }}
+              >
+                <ReactTooltip id="earnings" place="top" effect="solid">
+                  {tooltips.earnings}
+                </ReactTooltip>
+                <ReactSVG
+                  style={{
+                    paddingTop: 0,
+                  }}
+                  src={earningsIcon}
+                />
+                <Paragraph size="m-400" style={{ paddingBottom: 5 }}>
+                  &nbsp;Est. yearly yield&nbsp;
+                </Paragraph>
+                <ReactSVG
+                  data-tip
+                  data-for="earnings"
+                  src={errorOutlineIcon}
+                  wrapper="span"
+                />
+              </div>
+
+              <div
+                style={{
+                  minWidth: '100%',
+                  display: 'flex',
+                  flexDirection: 'row',
+                  justifyContent: 'space-around',
+                }}
+              >
+                <DisplayValue
+                  size="xl-400"
+                  color="#1ACD8C"
+                  value={totalReward}
+                />
+                <Paragraph size="xl-400" style={{ color: '#4A38AE' }}>
+                  &nbsp; / &nbsp;
+                </Paragraph>
+                <DisplayValue
+                  size="xl-400"
+                  color="#4A38AE"
+                  value={totalRewardInUsd}
+                  symbol="$"
+                />
+              </div>
+            </div>
+          </div>
+        </>
       </div>
     </div>
   );
 };
 InfoBlock.propTypes = {
+  poolsArr: PropTypes.array,
   account: PropTypes.string,
-  totalReward: PropTypes.any,
-  totalStaked: PropTypes.any,
-  totalRewardInUsd: PropTypes.any,
-  totalStakedInUsd: PropTypes.any,
 };
 export default React.memo(InfoBlock);
